And I can draw markers on a plot. Next are lines, then actual datasets as line plots. Then, filled line plots, then histiograms and bar plots. Et cetera.
Still need to get axes done, and more work on rt ui required.
Ultimately, must check and finish actual rt calculations.


git-svn-id: file:///mnt/svn/rtmath@56 1d52797c-35fc-48e8-8e80-2db72aa187e2

+++ rtmath/rtmath-base/polynomial.cpp
@@ -373,7 +373,7 @@ namespace rtmath {
 	void polynomial::truncate(unsigned int pow)
 	{
 		// Truncate at the specified power
-		unsigned int max = _coeffs.size();
+		unsigned int max = (unsigned int) _coeffs.size();
 		for (unsigned int i=pow+1; i<=max; i++)
 		{
 			// Safety check. If it exists, delete it
